+++ travel-app/src/components/MainPage/MainPage.tsx
@@ -1,18 +1,42 @@
-import React, { Component } from 'react';
+import React, { Component, PropsWithChildren } from 'react';
 import Header from '../Header/Header';
 import CountriesCards from './CountriesCards/CountriesCards';
+import CountryPage from '../country-page/CountryPage'
 import Footer from '../Footer/Footer';
-import './MainPage.css'
+import './MainPage.css';
+import {
+  Switch,
+  Route,
+  BrowserRouter
+} from "react-router-dom";
+
 
 export default class MainPage extends Component {
 
   render() {
-    return (
+    const counryPageProps = {
+      head: 'Germany',
+      titleImgSrc: 'https://smapse.ru/storage/2018/12/34133860-1854685924582132-4255627268174381056-n.jpg',
+      capital: 'Berlin',
+      videoSrc: 'https://www.youtube.com/watch?v=UK6x7pGBYnw&t=0s',
+      countryInfo: 'Lorem ipsum, dolor sit amet consectetur adipisicing elit. Dolores ipsam facilis harum! Quia eaque corrupti, eligendi atque perspiciatis architecto beatae quam officia perferendis modi, corporis eos voluptate qui nihil molestias.',
+    }
+    return ( 
       <div className='container'>
-        <Header />
-        <CountriesCards />
-        <Footer />
-      </div>
+        <BrowserRouter>
+          <Header />
+          <Switch>
+            <Route exact path="/">
+              <CountriesCards />
+            </Route>
+            <Route path="/country">
+              <CountryPage counryPageProps={counryPageProps} />
+            </Route>
+          </Switch>
+          <Footer />
+
+        </BrowserRouter>
+      </div >
     );
   }
 }
