+++ travel-app/src/components/country-page/video/video.css
@@ -5,5 +5,3 @@
     align-self: center;
     margin: 0 auto;
 }
-
-
