+++ travel-app/src/index.tsx
@@ -4,6 +4,7 @@ import 'semantic-ui-css/semantic.min.css'
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import 'semantic-ui-css/semantic.min.css'
+import './index.css';
 
 
 ReactDOM.render(
